backend/simsy/requirements.txt
Django
djangorestframework
django-cors-headers
django-rest-knox
python-dotenv
Pillow